fix(api): estimates endpoint layering and path containment
The estimates payload helper moves to analysis/estimates.py where all
its imports are intra-layer, and the scope filter moves to
analysis/manifest_scope.py (pure manifest logic; _cli_resolution
re-exports it for the CLI). The api route keeps a single lazy
analysis import, baselined like the existing _evaluation_routes seam,
instead of the six new layer violations CI flagged.

The route also resolves the project dir and checks it stays inside the
reports root inline, so CodeQL recognizes the path-injection barrier
for the two alerts it raised.
