+++ src/quodeq/_cli_resolution.py
@@ -20,6 +20,9 @@ from quodeq.shared.utils import is_repo_url, project_name_from_repo, read_json
 from quodeq.shared.validation import validate_path_segment
 from quodeq.analysis.manifest import SourceManifest, build_manifest, detect_language
 from quodeq.analysis.manifest_models import AnalysisTarget
+# Re-exported: moved to the analysis layer (pure manifest logic); CLI modules
+# keep their historical `from quodeq._cli_resolution import ...` path.
+from quodeq.analysis.manifest_scope import _filter_manifest_by_scope  # noqa: F401
 from quodeq.analysis.runner import load_universal_dimensions
 
 import logging
@@ -227,46 +230,6 @@ def _resolve_single_file(src: Path) -> tuple[Path, str | None]:
     return project_root, single_file
 
 
-def _filter_manifest_by_scope(
-    manifest: "SourceManifest | None", scope_path: str,
-) -> "SourceManifest | None":
-    """Narrow a manifest to only files under *scope_path*.
-
-    Returns None (with error printed) when no files match.
-    """
-    if not manifest or not manifest.targets:
-        return manifest
-
-    prefix = scope_path.rstrip("/") + "/"
-    scoped_targets: list[AnalysisTarget] = []
-    total = 0
-    all_stats: dict[str, int] = {}
-
-    for t in manifest.targets:
-        scoped_files = [f for f in t.source_files if f.startswith(prefix) or f == scope_path]
-        if not scoped_files:
-            continue
-        stats: dict[str, int] = {}
-        for f in scoped_files:
-            ext = os.path.splitext(f)[1]
-            if ext:
-                stats[ext] = stats.get(ext, 0) + 1
-        scoped_targets.append(AnalysisTarget(
-            name=t.name, language=t.language,
-            source_files=scoped_files, total_files=len(scoped_files),
-            language_stats=stats, category=t.category,
-        ))
-        total += len(scoped_files)
-        for k, v in stats.items():
-            all_stats[k] = all_stats.get(k, 0) + v
-
-    if scoped_targets:
-        print(f"Scope filter: {total} files under '{scope_path}'", file=sys.stderr)
-        return SourceManifest(targets=scoped_targets, total_files=total, language_stats=all_stats)
-
-    print(f"No source files found under scope '{scope_path}'", file=sys.stderr)
-    print("The scoped folder contains no recognized source code files.", file=sys.stderr)
-    return None
 
 
 def _override_manifest_single_file(
